pyproject.toml
[tool.poetry]
name = "bizol-ecommerce"
version = "0.1.0"
description = ""
authors = ["SabirovAsadbek8802 <sabirovasadbek932848802@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.3"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"

[tool.poetry.group.dev.dependencies]
Django = "^5.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
